updating the Appointment Controller

+++ src/main/java/com/ABCLab/ABCLab/Repository/AppointmentRepo.java
@@ -3,7 +3,9 @@ package com.ABCLab.ABCLab.Repository;
 import com.ABCLab.ABCLab.Model.Appointment;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface AppointmentRepo extends MongoRepository<Appointment, Integer> {
+import java.math.BigInteger;
+
+public interface AppointmentRepo extends MongoRepository<Appointment, BigInteger> {
